MAJOR CHANGES!!! REVIEW BEFORE MERGING!
Updated the models and put everything under the "model" package.

Every model is put in the package according to their swagger documentation (application, discover, friends, react, post and user).
The models that are used frequently (e.g. : Category, CheckIn, Dimension...), are placed in the "base" package.

Just the duplicate models are removed. the functionality is still the same. I've normally tested the app and then pushed the code in my branch.

Please review the code in my branch first before merging and let me know of any issues.